Create OpenDNS Diagnostic Tool.xml
Requires PowerShell to be installed on Agent.

This procedure can be run to execute the OpenDNS Diagnostic tool to provide diagnostics to OpenDNS support when opening a support ticket.  This procedure downloads the OpenDNS Diagnostic tool to the Agent Temp folder, prompts for a domain to use for testing (helpful if a specific domain is problematic) and saves the output text file of the results to the Agent Documents.  It also captures the URL with a copy of the results, which can be provided to OpenDNS Support.  The URL is emailed to the Default.AdminEmail.